backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.1.20
langchain-google-genai==1.0.1
langgraph==0.0.26
google-generativeai==0.4.1
pydantic==2.10.5
pydantic-settings==2.7.1
python-multipart==0.0.9
websockets==12.0
